+++ src/components/small/menu.js
@@ -1,6 +1,7 @@
 import 'semantic-ui-css/semantic.min.css';
 import React, { Component } from 'react'
 import { Button, Dropdown, Menu } from 'semantic-ui-react'
+import'../../App.css'
 
 export default class MenuExampleSizeLarge extends Component {
   state = { activeItem: 'home' }
@@ -13,6 +14,9 @@ export default class MenuExampleSizeLarge extends Component {
     return (
       <Menu  className="navigasiright">
         <Menu.Menu position='right'>
+          <Menu.Item className='menu'>
+             List Daftar Kelas
+          </Menu.Item>
           <Menu.Item>
              <Button color='teal'>Buat Kelas</Button>
           </Menu.Item>
